chore: move about some types
